12-drf-profiles-api/requirements.txt
django==2.2
djangorestframework==3.9.2
markdown
django-filter